# services/user_service/docs/api.yaml
openapi: 3.1.0
info:
  title: user_service
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
  /auth/login:
    post:
      tags:
      - auth
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
      - auth
      operationId: refresh_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/create:
    post:
      tags:
      - users
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
        required: true
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
      - users
      operationId: get_user_by_id
      parameters:
      - name: id
        in: path
        description: User's unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        password:
          type: string
        username:
          type: string
    LoginResponse:
      type: object
      required:
      - refresh_token
      - access_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    NotificationContext:
      type: object
      properties:
        email:
          type:
          - string
          - 'null'
        telegram:
          type:
          - string
          - 'null'
    RefreshTokenRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshTokenResponse:
      type: object
      required:
      - access_token
      properties:
        access_token:
          type: string
    User:
      type: object
      required:
      - id
      - username
      - notification_context
      properties:
        id:
          type: string
          format: uuid
        notification_context:
          $ref: '#/components/schemas/NotificationContext'
        username:
          type: string
    UserCreateRequest:
      type: object
      required:
      - username
      - notification_context
      - password
      properties:
        notification_context:
          $ref: '#/components/schemas/NotificationContext'
        password:
          type: string
        username:
          type: string
    UserCreatedResponse:
      type: object
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
tags:
- name: users
  description: User management endpoints
- name: auth
  description: Authentication endpoints